Sonarr-plugin/sonarr-plugin/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1
jinja2>=3.1.2
watchdog>=3.0.0
python-multipart>=0.0.6
pyinstaller>=6.15.0 
pystray>=0.19.0
Pillow>=10.0.0
packaging>=23.0 